+++ frontend/src/home_and_profile/Home.tsx
@@ -1,9 +1,6 @@
-import Timeline from './Timeline.tsx';
+import ViewPosts from './ViewPosts.tsx';
 import CreatePost from './CreatePost.tsx';
-import './Home.css'
-import { Link } from 'react-router-dom';
 import SideBar from './SideBar.tsx';
-import { Button } from '@/components/ui/Button.tsx';
 import { useCookies } from 'react-cookie';
 
 export default function Home() {  
@@ -15,24 +12,9 @@ export default function Home() {
                 <SideBar />
             </div>
             
-            <div className='flex-initial basis-2/3 outline'>
-                <div className="bg-green-800 w-70 h-1000 text-center">
-                    <CreatePost createrId = { cookies.myId /*自分のid*/ } />
-                    <Timeline />
-                </div>
-
-                <div className='box'>
-                    <span className='box-title'>ここにタイトル</span>
-                    <p>ここに文章。タイトルのところに長い文章を書くとスタイルが崩れるらしい。最大何文字なのか、どうやって対応していけば良いのかは後で考えれば良いのかな？それともそもそも枠なんて必要ないのかしら？</p>
-                </div>
-
-                <h2 className='homeH2'>以下、テスト用リンク</h2>
-                <Link to="/profile/0">id=0のプロフィール画面</Link>
-
-                <div className="bg-neutral-50 dark:bg-white">
-                    aaaaa
-                </div>
-                <Button>test用クリック</Button>
+            <div className='flex-initial basis-2/3 m-5'>
+                <CreatePost createrId = { cookies.myId } />
+                <ViewPosts />
             </div>
         </div>
     )
